SDL2: Fix SDL_TEXTINPUT handling
The previous code was only examining the first character of the string
from the SDL_TEXTINPUT event, although SDL_TEXTINPUT can return multiple
characters at a time. For example, if I enter "玉米超人“ with my chinese
IME, in wide mode, only the first character will be evaluated.

Depending on the IME, it may be possible in non-wide mode to also
receive multiple characters at a time. Thus, when we get a SDL_TEXTINPUT
event, we should process each character before polling for the next event.